Added support for multiple index buffers in vertex buffer class. Added a cylinder form. Wrote a terrain chunk.
git-svn-id: svn://svn.code.sf.net/p/captainsengine/code/trunk@232 dc34a4ac-3e42-4537-918b-2dea8b01751e

+++ Engine/renderer/vertexbuffer.cpp
@@ -10,7 +10,7 @@ VertexBuffer::~VertexBuffer(){
 
 }
 
-void VertexBuffer::create(int type, int vertexBufferSize, int indexBufferSize){
+void VertexBuffer::create(int type, int vertexBufferSize, short indexBufferSize){
 
 }
